Refactor log transaction
