+++ src/plugins/explorer/index.js
@@ -1,14 +1,14 @@
 'use strict'
 
 const debug = require('debug')('met-wallet:core:explorer')
-const pDefer = require('p-defer')
 const Web3 = require('web3')
 
 const createEventsRegistry = require('./events')
+const createIndexer = require('./indexer')
+const createLogTransaction = require('./log-transaction')
 const createQueue = require('./queue')
 const createStream = require('./blocks-stream')
 const createTransactionSyncer = require('./sync-transactions')
-const createIndexer = require('./indexer')
 const refreshTransaction = require('./refresh-transactions')
 const tryParseEventLog = require('./parse-log')
 
@@ -51,39 +51,9 @@ function createPlugin () {
       })
     })
 
-    function logTransaction (promiEvent, from, meta) {
-      // PromiEvent objects shall be wrapped to avoid the promise chain to
-      // cast it to a plain promise
-      if (promiEvent.once) {
-        const deferred = pDefer()
-
-        promiEvent.once('transactionHash', function (hash) {
-          queue.addTransaction(from, meta)(hash)
-        })
-        promiEvent.once('receipt', function (receipt) {
-          queue.addTransaction(from)(receipt.transactionHash)
-          deferred.resolve({ receipt })
-        })
-        promiEvent.once('error', function (err) {
-          promiEvent.removeAllListeners()
-          deferred.reject(err)
-        })
-
-        return deferred.promise
-      }
-
-      // This is not a wrapped PromiEvent object. It shall be a plain promise
-      // instead.
-      const promise = promiEvent
-      return promise.then(function (receipt) {
-        queue.addTransaction(from)(receipt.transactionHash)
-        return { receipt }
-      })
-    }
-
     return {
       api: {
-        logTransaction,
+        logTransaction: createLogTransaction(queue),
         refreshAllTransactions: syncer.refreshAllTransactions,
         refreshTransaction: refreshTransaction(web3, eventsRegistry, queue),
         registerEvent: eventsRegistry.register,
